Plot smart charging results

HTML interactive plot showing the comparison between the smart charging setpoint and the actual EV demand after the smart charging program. Also, it is possible to plot the original EV demand.

Usage

plot_smart_charging(smart_charging, sessions = NULL, ...)

Arguments

smart_charging

SmartCharging object, returned by function smart_charging()

sessions

tibble, sessions data set containig the following variables: "Session", "Timecycle", "Profile", "ConnectionStartDateTime", "ConnectionHours", "Power" and "Energy"

...

extra arguments to pass to dygraphs::dyOptions function

Value

dygraphs plot

Examples

library(dplyr)
sessions <- evsim::california_ev_sessions_profiles %>%
  slice_head(n = 50) %>%
  evsim::adapt_charging_features(time_resolution = 15)
sessions_demand <- evsim::get_demand(sessions, resolution = 15)

# Don't require any other variable than datetime, since we don't
# care about local generation (just peak shaving objective)
opt_data <- tibble(
  datetime = sessions_demand$datetime,
  production = 0
)

sc_results <- smart_charging(
  sessions, opt_data,
  opt_objective = "grid",
  method = "curtail",
  window_days = 1, window_start_hour = 6
)

# Plot of setpoint and final EV demand
plot_smart_charging(sc_results)

# Plot with original demand
plot_smart_charging(sc_results, sessions = sessions)

# Native `plot` function also works
plot(sc_results)