Schedule sessions according to optimal setpoint

Schedule sessions according to optimal setpoint

Usage

schedule_sessions(
  sessions,
  setpoint,
  method,
  power_th = 0,
  charging_power_min = 0.5,
  include_log = FALSE,
  show_progress = TRUE
)

Arguments

sessions

tibble, sessions data set normalized from normalize_sessions function

setpoint

tibble with columns datetime, timeslot and rest of columns being the names of user profiles

method

character, being postpone, curtail or interrupt.

power_th

numeric, power threshold (between 0 and 1) accepted from setpoint. For example, with power_th = 0.1 and setpoint = 100 for a certain time slot, then sessions' demand can reach a value of 110 without needing to schedule sessions.

charging_power_min

numeric, minimum allowed ratio (between 0 and 1) of nominal power. For example, if charging_power_min = 0.5 and method = 'curtail', sessions' charging power can only be curtailed until the 50% of the nominal charging power (i.e. Power variable in sessions tibble).

include_log

logical, whether to output the algorithm messages for every user profile and time-slot

show_progress

logical, whether to output the progress bar in the console

Value

list of two elements sessions and log