Get a summary of the energy charged per session

A table is provided containing the originally required energy for every session and the actual energy charged with the smart charging program (also in percentage).

Usage

summarise_energy_charged(smart_charging, sessions)

Arguments

smart_charging

SmartCharging object, returned by function smart_charging()

sessions

tibble, sessions data set containig the following variables: "Session", "Timecycle", "Profile", "ConnectionStartDateTime", "ConnectionHours", "Power" and "Energy".

Value

tibble

Examples

library(dplyr)

sessions <- evsim::california_ev_sessions_profiles %>%
  slice_head(n = 50) %>%
  evsim::adapt_charging_features(time_resolution = 15)
sessions_demand <- evsim::get_demand(sessions, resolution = 15)

# Don't require any other variable than datetime, since we don't
# care about local generation (just peak shaving objective)
opt_data <- tibble(
  datetime = sessions_demand$datetime,
  production = 0
)
sc_results <- smart_charging(
  sessions, opt_data, opt_objective = "grid", method = "curtail",
  window_days = 1, window_start_hour = 6,
  responsive = list(Workday = list(Worktime = 0.9)),
  energy_min = 0.5
)

summarise_energy_charged(sc_results, sessions)
#> # A tibble: 50 × 4
#>    Session EnergyCharged EnergyRequired PctEnergyCharged
#>    <chr>           <dbl>          <dbl>            <dbl>
#>  1 S1               6.31           6.31              100
#>  2 S2              11.4           22.8                50
#>  3 S3              12.7           25.4                50
#>  4 S4               7.53           7.53              100
#>  5 S5              10.3           20.7                50
#>  6 S6               2.51           5.01               50
#>  7 S7               4.41           8.42               52
#>  8 S8               8.86          17.1                52
#>  9 S9               7.85          15.7                50
#> 10 S10             58.9           58.9               100
#> # ℹ 40 more rows